requirements.txt
streamlit == 1.45.1
st-gsheets-connection
google-genai
streamlit-authenticator == 0.4.2
bcrypt